ci(all): remove testing for Magento v2.3 (pre-2.3.7).
It turns out that there's some nasty code in Magento that makes some assumptions about Xdebug that aren't really safe to make. When Github updated their VE to support Xdebug 3.0, this caused tests to break without the tests actually being broken. This SHOULD be fixed in v2.3.7 of Magento, but it isn't out yet, so I'm removing the tests.

https://github.com/magento/magento2/issues/28961